pyproject.toml
[project]
name = "fasthtml-test-faststrap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faststrap>=0.3.0",
    "python-fasthtml>=0.12.36",
]
